# arm64 crc optimise