--- a/util/crc32c_arm64.cc
+++ b/util/crc32c_arm64.cc
@@ -44,8 +44,8 @@ uint32_t crc32c_arm64(uint32_t crc, unsigned char const *data,
      */
     uint32_t k0 = 0xe417f38a, k1 = 0x8f158014;
 
-	/* Prefetch data for following block to avoid cache miss */
-	PREF1KL1((uint8_t *)buf64, 1024);
+    /* Prefetch data for following block to avoid cache miss */
+    PREF1KL1((uint8_t *)buf64, 1024);
 	
     /* First 8 byte for better pipelining */
     crc0 = crc32c_u64(crc, *buf64++);